requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
gunicorn==21.2.0
Jinja2==3.1.3
Werkzeug==2.3.7
itsdangerous==2.1.2
email-validator==2.1.1
bcrypt==4.1.2
Flask-WTF==1.2.1
wtforms[email]==3.1.2
Flask-Bootstrap==3.3.7.1
Pillow==10.2.0

